+++ packages/storefront-next-runtime/CHANGELOG.md
@@ -4,3 +4,4 @@
 - Add support to MRT Data Layer access [#1215](https://github.com/commerce-emu/storefront-next/pull/1215)
 - Add currency detection to site-context middleware (@W-21787262) ([#1342](https://github.com/commerce-emu/storefront-next/pull/1342))
 - Add runtime support for custom API client typing and proxy client composition (@W-21549425)
+- Add support for OOTB API Key for Google Address Autocomplete feature from MRT Data Layer (@W-22130944) ([#1509](https://github.com/commerce-emu/storefront-next/pull/1509))